fix ttl in logical collection

+++ arangod/VocBase/LogicalCollection.h
@@ -392,7 +392,7 @@ class LogicalCollection {
   mutable basics::ReadWriteLock _infoLock;  // lock protecting the info
 
   std::vector<std::pair<std::string, double>> _clusterEstimates;
-  uint64_t _clusterEstimateTTL; //only valid if above vector is not empty
+  double _clusterEstimateTTL; //only valid if above vector is not empty
   basics::ReadWriteLock _clusterEstimatesLock;
 };
